+++ CHANGELOG.md
@@ -14,6 +14,7 @@ New features:
 - XIT FINBS
 - XIT HQUC - HQ Upgrade Calculator
 - XIT REP now shows a weight/volume/cost table
+- Building list with building count in BS buffer
 
 Changes:
 - XIT command renames: // TODO